refactor(cocoapods): Move Podspec to a dedicated file
Separate the CocoaPods model from the analyzer.
