+++ plugins/package-managers/cocoapods/src/main/kotlin/CocoaPods.kt
@@ -19,12 +19,6 @@
 
 package org.ossreviewtoolkit.plugins.packagemanagers.cocoapods
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-import com.fasterxml.jackson.core.JsonParser
-import com.fasterxml.jackson.databind.DeserializationContext
-import com.fasterxml.jackson.databind.JsonNode
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize
-import com.fasterxml.jackson.databind.deser.std.StdDeserializer
 import com.fasterxml.jackson.databind.node.ObjectNode
 
 import java.io.File
@@ -316,48 +310,4 @@ private fun parseLockfile(podfileLock: File): LockfileData {
     }
 
     return LockfileData(dependencies, externalSources.orEmpty())
-}
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-private data class Podspec(
-    val name: String = "",
-    val version: String = "",
-    @JsonDeserialize(using = LicenseDeserializer::class)
-    val license: String = "",
-    val summary: String = "",
-    val homepage: String = "",
-    val source: Map<String, String> = emptyMap(),
-    private val subspecs: List<Podspec> = emptyList()
-) {
-    fun withSubspecs(): List<Podspec> {
-        val result = mutableListOf<Podspec>()
-
-        fun add(spec: Podspec, namePrefix: String) {
-            val name = "$namePrefix${spec.name}"
-            result += copy(name = "$namePrefix${spec.name}")
-            spec.subspecs.forEach { add(it, "$name/") }
-        }
-
-        add(this, "")
-
-        return result
-    }
-}
-
-/**
- * Handle deserialization of the following two possible representations:
- *
- * 1. https://github.com/CocoaPods/Specs/blob/f75c24e7e9df1dac6ffa410a6fb30f01e026d4d6/Specs/8/5/e/SocketIOKit/2.0.1/SocketIOKit.podspec.json#L6-L9
- * 2. https://github.com/CocoaPods/Specs/blob/f75c24e7e9df1dac6ffa410a6fb30f01e026d4d6/Specs/8/5/e/FirebaseObjects/0.0.1/FirebaseObjects.podspec.json#L6
- */
-private class LicenseDeserializer : StdDeserializer<String>(String::class.java) {
-    override fun deserialize(parser: JsonParser, context: DeserializationContext): String {
-        val node = parser.codec.readTree<JsonNode>(parser)
-
-        return if (node.isTextual) {
-            node.textValue()
-        } else {
-            node["type"].textValueOrEmpty()
-        }
-    }
 }
